+++ bot/commands/info/source.js
@@ -1,5 +1,6 @@
 'use strict';
 var repo = require(process.cwd()+'/repo');
+var moment = require('moment');
 
 function findTrigger(bot, db, data) {
   if (data.params === void(0) || data.params.length < 1) {
@@ -22,9 +23,28 @@ function findTrigger(bot, db, data) {
       var keys = Object.keys(val);
       var result = val[keys[0]];
       var theAuthor = result.Author || "unknown";
-      var status = result.status || "created/updated";
+      let extendedInfo = false;
+      let chatMsg = `the trigger ${trigger}`;
+      let chatMsgEnd = ` was created/updated by ${theAuthor}`;
 
-      return bot.sendChat(`the trigger ${trigger} was ${status} by ${theAuthor}`);
+      if (result.createdOn && result.createdBy) {
+        let when = moment(result.createdOn).format('MMM Do YYYY');
+        chatMsg += ` was created by ${result.createdBy} on ${when}`;
+        extendedInfo = true;
+      }
+
+      if (result.lastUpdated && result.status && result.status === 'updated') {
+        let when = moment(result.lastUpdated).format('MMM Do YYYY');
+        chatMsg += ` and was last updated by ${theAuthor} on ${when}`;
+        extendedInfo = true;
+      }
+
+      if (extendedInfo) {
+        return bot.sendChat(chatMsg);
+      } else {
+        return bot.sendChat(chatMsg + chatMsgEnd);
+      }
+
     }
   });
 }