+++ bot/events/connected.js
@@ -3,6 +3,11 @@ var mediaStore = require(process.cwd()+ '/bot/store/mediaInfo.js');
 var historyStore = require(process.cwd()+ '/bot/store/history.js');
 var repo = require(process.cwd()+'/repo');
 
+var _env = process.env.ENV;
+if (_env === null || typeof _env === 'undefined' || _env === '') {
+  _env = 'dev';
+}
+
 module.exports = function(bot, db) {
   bot.on('connected', function(data) {
     bot.log('info', 'BOT', 'Connected to ' + data);
@@ -30,6 +35,16 @@ module.exports = function(bot, db) {
       }
 
       historyStore.init(bot);
+
+      // store user info locally
+      var user = db.ref(_env + '/users');
+      user.on('value', function(snapshot){
+          var val = snapshot.val();
+          bot.chilloutUsers = val;
+        }, function(error){
+          bot.log('error', 'BOT', 'error getting users from firebase');
+      });
+
     }, 5000);
   });
 };